fix param

+++ net-adapter/src/endpoint_def.hpp
@@ -244,10 +244,7 @@ EndPointNode<E>::EndPointNode() :
                 *this,
                 CONNECTION_PARAM_NAME,
                 DEFAULT_CONNECTION_HOSTNAME),
-            util::declare_and_get_param<double>(
-                *this,
-                "net_batch_size",
-                65535)))},
+            util::declare_and_get_param<int>(*this, "net_batch_size", 65535)))},
     delay_queue{std::chrono::duration<double>{
         util::declare_and_get_param(*this, "net_delay_s", 0.0)}},
