+++ net-adapter/src/util/zenoh_utils.hpp
@@ -49,18 +49,22 @@ namespace util
 
 namespace zenoh_aliases
 {
-    using ZenohPub = ::zenoh::Publisher;
-    using ZenohSub = ::zenoh::Subscriber<void>;
-};
+using ZenohPub = ::zenoh::Publisher;
+using ZenohSub = ::zenoh::Subscriber<void>;
+};  // namespace zenoh_aliases
 
-inline zenoh::Config configDirectConnectTo(const std::string& hostname, uint16_t batch_size = 65535)
+inline zenoh::Config configDirectConnectTo(
+    const std::string& hostname,
+    int batch_size = 65535)
 {
     zenoh::Config config = zenoh::Config::create_default();
     config.insert_json5("mode", "\"peer\"");
     config.insert_json5("scouting/multicast/enabled", "false");
     config.insert_json5("listen/endpoints", "[\"tcp/0.0.0.0:7447\"]");
     config.insert_json5("connect/endpoints", "[\"tcp/" + hostname + ":7447\"]");
-    config.insert_json5("transport/link/tx/batch_size", std::to_string(batch_size));
+    config.insert_json5(
+        "transport/link/tx/batch_size",
+        std::to_string(batch_size));
     return config;
 }
